Adding super route presenter to app route presenter and autofill route presenter. Separating generic methods into parent.
